util: Add tabbing restriction module for improved accessibility
This commit adds a new JavaScript module that provides the functions
`restrictTabbing` and `removeTabbingRestriction`. These functions can
be used to restrict the tabbing focus within a specific container element,
which can improve accessibility for users who rely on keyboard navigation.
The module is added to the `core/js` directory, and can be imported by any
app that needs to implement tabbing restrictions.

Resolves : https://github.com/nextcloud/server/issues/36989

Signed-off-by: fenn-cs <fenn25.fn@gmail.com>